+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.2.2 (unreleased)
+0.3.0 (unreleased)
 ------------------
 
 * Add unload_file callback in TFTPUDPHandler, called when the transfer of a